requirements.txt
imageai==2.1.6
tensorflow==2.4.0
tensorflow-gpu==2.4.0
keras==2.4.3
numpy==1.19.3
Pillow==7.0.0
scipy==1.4.1
h5py==2.10.0
librosa==0.9.2
matplotlib==3.3.2
opencv-python==4.6.0.66
keras-resnet==0.2.0
praat-textgrids==1.3.1
lxml==4.9.1
scikit-image==0.19.3
yt-dlp==2022.7.18